+++ src/internal/transactionSubmission.ts
@@ -100,10 +100,10 @@ export async function buildTransactionPayload(
   args: { aptosConfig: AptosConfig } & InputGenerateTransactionData,
 ): Promise<AnyTransactionPayloadInstance> {
   const { aptosConfig, data } = args;
-
   // Merge in aptosConfig for remote ABI on non-script payloads
   let generateTransactionPayloadData: InputGenerateTransactionPayloadDataWithRemoteABI;
   let payload: AnyTransactionPayloadInstance;
+
   if ("bytecode" in data) {
     // TODO: Add ABI checking later
     payload = await generateTransactionPayload(data);
@@ -114,6 +114,7 @@ export async function buildTransactionPayload(
       function: data.function,
       functionArguments: data.functionArguments,
       typeArguments: data.typeArguments,
+      abi: data.abi,
     };
     payload = await generateTransactionPayload(generateTransactionPayloadData);
   } else {
@@ -122,6 +123,7 @@ export async function buildTransactionPayload(
       function: data.function,
       functionArguments: data.functionArguments,
       typeArguments: data.typeArguments,
+      abi: data.abi,
     };
     payload = await generateTransactionPayload(generateTransactionPayloadData);
   }